refactor story list to support multiple fragment types (#1052)
